requirements.txt
streamlit>=1.28
openai>=1.0
yt-dlp>=2024.1
moviepy>=1.0.3
python-dotenv>=1.0
pydub>=0.25.1
imageio-ffmpeg>=0.4.9
